refactor: 타입, api 함수 위치 변경
