+++ src/components/molecules/FormTextInput/index.stories.tsx
@@ -5,7 +5,7 @@ import FormControl from '../FormControl/FormControl';
 import { BorderBox } from '~/components/atoms/BorderBox';
 import { Button } from '~/components/atoms/Button';
 import { FormLabel } from '~/components/atoms/FormLabel';
-import { CreateEvent } from '~/services/eventService';
+import { CreateEvent } from '~/types/event';
 
 const meta = {
   title: 'Resumeme/Components/FormTextInput',
