+++ src/types/event.ts
@@ -1,0 +1,14 @@
+export type CreateEvent = {
+  info: {
+    title: string;
+    content: string;
+    maximumAttendee: number;
+  };
+  time: {
+    now: string;
+    openDateTime: string;
+    closeDateTime: string;
+    endDate: string;
+  };
+  positions: string[];
+};
